requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
python-decouple>=3.8
psycopg2-binary>=2.9
Pillow>=10.0
django-filter>=23.0
django-ratelimit>=4.1.0